+++ src/main/java/pretzel/dreamketcherbe/auth/jwt/TokenProvider.java
@@ -1,4 +1,4 @@
-package pretzel.dreamketcherbe.auth.utils;
+package pretzel.dreamketcherbe.auth.jwt;
 
 
 import pretzel.dreamketcherbe.auth.dto.AuthPayload;
