+++ src/main/java/pretzel/dreamketcherbe/auth/service/AuthService.java
@@ -1,6 +1,5 @@
 package pretzel.dreamketcherbe.auth.service;
 
-import ch.qos.logback.core.subst.Token;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -9,7 +8,7 @@ import pretzel.dreamketcherbe.auth.dto.AuthPayload;
 import pretzel.dreamketcherbe.auth.dto.TokenResponse;
 import pretzel.dreamketcherbe.auth.google.GoogleOAuthClient;
 import pretzel.dreamketcherbe.auth.google.dto.GoogleUserInfo;
-import pretzel.dreamketcherbe.auth.utils.TokenProvider;
+import pretzel.dreamketcherbe.auth.jwt.TokenProvider;
 import pretzel.dreamketcherbe.member.entity.Member;
 import pretzel.dreamketcherbe.member.repository.MemberRepository;
